+++ open-media-match/src/OpenMediaMatch/blueprints/ui.py
@@ -1,11 +1,12 @@
 import flask
-import os
-import requests
 
 from flask import Blueprint, abort
-from flask import request, redirect, url_for
+from flask import request, redirect
 
 from OpenMediaMatch.blueprints import matching, curation, hashing
+from OpenMediaMatch.persistence import get_storage
+from OpenMediaMatch.background_tasks import build_index
+from OpenMediaMatch.database import db
 
 bp = Blueprint("ui", __name__)
 
@@ -32,6 +33,29 @@ def upload():
     }
 
     return {"hashes": signals, "banks": sorted(banks)}
+
+
+@bp.route("/rebuild_index", methods=["POST"])
+def rebuild_index():
+    st_name = request.form.get("signal_type")
+    storage = get_storage()
+    if st_name is not None:
+        st = storage.get_signal_type_configs().get(st_name)
+        if st is None:
+            abort(404, f"No such signal type '{st_name}'")
+        if not st.enabled:
+            abort(400, f"Signal type {st_name} is disabled")
+        build_index.build_index(st.signal_type)
+        return {}
+    build_index.build_all_indices(storage, storage, storage)
+    return redirect("/")
+
+
+@bp.route("/factory_reset", methods=["POST"])
+def factory_reset():
+    db.drop_all()
+    db.create_all()
+    return redirect("/")
 
 
 @bp.route("/addcontent", methods=["POST"])
